LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-normalization/positionvalue-normalization.tentative-expected.txt

CONSOLE MESSAGE: TypeError: CSS.percent is not a function. (In 'CSS.percent(0)', 'CSS.percent' is undefined)

Harness Error (FAIL), message = TypeError: CSS.percent is not a function. (In 'CSS.percent(0)', 'CSS.percent' is undefined)
